fix the Merge Sort test on Hackerrank.

+++ src/com/amazon/MergeSort2.java
@@ -4,33 +4,34 @@ import java.util.Scanner;
 
 /**
  * Created by kaibohao on 2016-11-30.
+ * For test: https://www.hackerrank.com/challenges/ctci-merge-sort
  */
 public class MergeSort2 {
-    public static int sort(int[] arr) {
+    public static long sort(int[] arr) {
         if (arr.length == 0) {
             return 0;
         }
         return sort(arr, new int[arr.length], 0, arr.length - 1);
     }
 
-    private static int sort(int[] arr, int[] temp, int low, int high) {
+    private static long sort(int[] arr, int[] temp, int low, int high) {
         if (low >= high)
             return 0;
         int mid = (low + high) / 2;
-        int leftTimes = sort(arr, temp, low, mid);
-        int rightTimes = sort(arr, temp, mid + 1, high);
-        int mergedTimes = merge(arr, temp, low, mid, high);
+        long leftTimes = sort(arr, temp, low, mid);
+        long rightTimes = sort(arr, temp, mid + 1, high);
+        long mergedTimes = merge(arr, temp, low, mid, high);
         return leftTimes + rightTimes + mergedTimes;
     }
 
-    private static int merge(int[] arr, int[] temp, int low, int mid, int high) {
+    private static long merge(int[] arr, int[] temp, int low, int mid, int high) {
         for (int i = low; i <= high; i++) {
             temp[i] = arr[i];
         }
         int current = low;
         int left = low;
         int right = mid + 1;
-        int swapTimes = 0;
+        long swapTimes = 0;
         while (left <= mid && right <= high) {
             if (temp[left] <= temp[right]) {
                 arr[current] = temp[left];
@@ -58,7 +59,7 @@ public class MergeSort2 {
             for(int arr_i=0; arr_i < n; arr_i++){
                 arr[arr_i] = in.nextInt();
             }
-            int swaps = MergeSort2.sort(arr);
+            long swaps = MergeSort2.sort(arr);
             System.out.println(swaps);
         }
     }
